Prevent EntropyMixer from registering with service manager.
EntropyMixer is designed to be triggered in two ways, as either the
result of a self-set timer or of one of three system broadcasts.  It is not
meant to be triggered in any other way, so exposing it via servicemanager
is undesirable.

Bug: 18106000

Cherry-pick of commit: 7d7e2be3491ef61d2a13acf50b8b6a0ab62c5f56

Change-Id: I9aeb35e7ffde75090f4234ea193514fb883b1425