Support connecting to networks with misconfigured subnet masks.
In K and earlier, we would connect to a network where the gateway
was not covered by the subnet mask of the IP address. This is an
invalid configuration, but it used to work, and other OSes appear
to accept it too, so support it.

Bug: 19067207

(cherry picked from commit d7dd03fab712017bae8e5f7057e3aa09de0a56d8)

Change-Id: I80088f291466dbd5a47f360dcc1620acee5cf57e