Recalculate input method top edge.
The bottom edge of an input method target window is set using the top
edge of the input method window. A switch from using the input
method's content frame's top edge to its display frame's top edge
(ag/634725) caused the underlying window to be obscured. This fix uses
the lower of the two frames to set the bottom edge of the target
window.

Fixes bug 19434816.

Change-Id: I5fe73fbe8a432b24a12c030f029cdfcd55e23274